Reran `bulk_training.py`, imported `output_model_different_random_seeds` module, reran and works.
Have many output and graph files generated by older code. They are still good but are not generated by this `bulk_training.py`. Rather some were output from the bulk training of the Jupyter notebook (only said it is obsolete because I am not updating it with new code)

In the next run and commit, I will delete all output and graph files and will generate again possibly with increasing the range of random seeds to `6`. This is to reduce clutter and fully migrate to the new file structure.